+++ mex/editor/search/models.py
@@ -1,0 +1,9 @@
+import reflex as rx
+
+
+class SearchResult(rx.Base):
+    """Search result preview."""
+
+    identifier: str
+    title: str
+    preview: str
